refactor(refs T30969): move DpCard to components root
