+++ src/components/core/index.js
@@ -2,7 +2,6 @@ import dataTableSearch from './DpDataTable/DataTableSearch'
 import DpAnonymizeText from './DpAnonymizeText'
 import DpAutocomplete from './DpAutocomplete'
 import DpBulkEditHeader from './DpBulkEditHeader'
-import DpCard from './DpCard'
 import DpChangeStateAtDate from './DpChangeStateAtDate'
 import DpColumnSelector from './DpDataTable/DpColumnSelector'
 import DpCopyPasteButton from './DpCopyPasteButton'
@@ -36,7 +35,6 @@ export {
   DpAnonymizeText,
   DpAutocomplete,
   DpBulkEditHeader,
-  DpCard,
   DpChangeStateAtDate,
   DpColumnSelector,
   DpCopyPasteButton,
